+++ src/engine/physics2d/Force.java
@@ -1,5 +1,5 @@
 package engine.physics2d; 
 
 public class Force {
-	public final static float GRAVITY = 1.8f;
+	public final static double GRAVITY = 1.8;
 }
